+++ backend/src/main/java/minionz/backend/chat/message/model/MessageStatus.java
@@ -1,4 +1,4 @@
-package minionz.backend.message.model;
+package minionz.backend.chat.message.model;
 
 public enum MessageStatus {
     UNREAD,
